requirements.txt
alembic==1.4.3
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.1.0
